move c++ binding to new board info getter

+++ src/board_controller/synthetic_board.cpp
@@ -131,7 +131,7 @@ void SyntheticBoard::read_thread ()
         base_wave[i] = this->amplitude * sin (1.8f * i * rads + this->shift);
     }
     // eeg channels + 3 accel channels + package num
-    float *package = new float[this->num_channels + 3 + 1];
+    double *package = new double[this->num_channels + 3 + 1];
     // random distr for noise
     std::random_device rd;
     std::mt19937 mt (rd ());
@@ -145,7 +145,7 @@ void SyntheticBoard::read_thread ()
     while (this->keep_alive)
     {
         // package num
-        package[0] = (float)counter;
+        package[0] = (double)counter;
         // eeg
         for (int i = 0; i < this->num_channels; i++)
         {
